security ok

+++ src/main/java/com/gestion/gestionecole/controller/AnneeScolaireController.java
@@ -84,7 +84,7 @@ public class AnneeScolaireController {
 	}
 	
 	@GetMapping("/countAnneeScolaire")
-	public Integer countAnneeScolaire() {
+	public Long countAnneeScolaire() {
 		return anneeScolaireService.count();
 	}
